Implemented base_image_checker.py
Validates if required files exist and have correct permissions
Validates data authenticity of base-image tarball

Logging to be implemented in the future.
nspawn.org public key info is hard-coded as a global variable - ok for now but must review in the future.
Might change exit codes and subprocess error hadndling depending on how this will be implemented